# src/main/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flowrider API
  description: |
    REST API for Flowrider - AI Session Orchestration Platform

    Flowrider is an Electron app for orchestrating multiple AI terminal sessions using a 3D icosahedron interface (20 faces = 20 sessions).

    ## Authentication
    All API requests require authentication using an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per API key.

    ## Error Handling
    The API uses standard HTTP status codes and returns error details in JSON format.
  version: 1.0.0
  contact:
    name: Josephas LLC
    url: https://github.com/josephas-llc/flowrider
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3737/api
    description: Local development server
  - url: https://api.flowrider.local/api
    description: Production server (when deployed)
...
security:
  - ApiKeyAuth: []
paths:
...
  /sessions:
    get:
      tags:
        - Sessions
      summary: List all sessions
      description: Retrieve a list of all AI terminal sessions
      operationId: listSessions
      parameters:
        - name: status
          in: query
          description: Filter sessions by status
          schema:
            type: string
            enum: [active, idle, terminated]
        - name: limit
          in: query
          description: Maximum number of sessions to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of sessions to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  total:
                    type: integer
                    description: Total number of sessions
                    example: 42
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Sessions
      summary: Create a new session
      description: Create a new AI terminal session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - faceIndex
                - workingDir
              properties:
                name:
                  type: string
                  description: Session name
                  example: "Backend API Development"
                faceIndex:
                  type: integer
                  description: Icosahedron face index (0-19)
                  minimum: 0
                  maximum: 19
                  example: 5
                workingDir:
                  type: string
                  description: Working directory path
                  example: "/Users/dev/my-project"
                projectId:
                  type: string
                  description: Associated project ID
                  example: "proj_123abc"
                aiProvider:
                  type: string
                  description: AI provider to use
                  enum: [ollama, claude, openai, gemini]
                  example: "claude"
                model:
                  type: string
                  description: AI model name
                  example: "claude-opus-4-5"
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Session limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/{id}:
    get:
      tags:
        - Sessions
      summary: Get session details
      description: Retrieve details for a specific session
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Sessions
      summary: Update session
      description: Update session details
      operationId: updateSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Session Name"
                status:
                  type: string
                  enum: [active, idle, terminated]
                  example: "active"
                currentTask:
                  type: string
                  example: "Debugging authentication flow"
                tags:
                  type: array
                  items:
                    type: string
                  example: ["backend", "api", "typescript"]
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Sessions
      summary: Delete session
      description: Terminate and delete a session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Session deleted successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /webhooks:
    get:
      tags:
        - Webhooks
      summary: List all webhooks
      description: Retrieve a list of all configured webhooks
      operationId: listWebhooks
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Webhooks
      summary: Create a new webhook
      description: Register a new webhook endpoint
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                  description: Webhook endpoint URL
                  example: "https://api.example.com/webhooks/flowrider"
                events:
                  type: array
                  description: Events to subscribe to
                  items:
                    type: string
                    enum:
                      - session.created
                      - session.updated
                      - session.deleted
                      - session.output
                      - project.created
                      - project.updated
                      - project.deleted
                      - deployment.started
                      - deployment.completed
                      - deployment.failed
                  example: ["session.created", "session.output"]
                secret:
                  type: string
                  description: Secret for webhook signature verification
                  example: "whsec_abc123def456"
                active:
                  type: boolean
                  default: true
                  example: true
      responses:
        '201':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /webhooks/{id}:
    delete:
      tags:
        - Webhooks
      summary: Delete webhook
      description: Remove a webhook endpoint
      operationId: deleteWebhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Webhook deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Webhook deleted successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: Session ID
      schema:
        type: string
        example: "face-5"
...
    WebhookId:
      name: id
      in: path
      required: true
      description: Webhook ID
      schema:
        type: string
        example: "wh_456def"
...
  schemas:
    Session:
      type: object
      properties:
        id:
          type: string
          example: "face-5"
        name:
          type: string
          example: "Backend API Development"
        faceIndex:
          type: integer
          minimum: 0
          maximum: 19
          example: 5
        status:
          type: string
          enum: [active, idle, terminated]
          example: "active"
        workingDir:
          type: string
          example: "/Users/dev/my-project"
        projectId:
          type: string
          nullable: true
          example: "proj_123abc"
        tmuxSessionName:
          type: string
          example: "flowrider-face-5"
        aiProvider:
          type: string
          enum: [ollama, claude, openai, gemini]
          example: "claude"
        model:
          type: string
          example: "claude-opus-4-5"
        currentTask:
          type: string
          nullable: true
          example: "Implementing authentication middleware"
        tags:
          type: array
          items:
            type: string
          example: ["backend", "api", "typescript"]
        createdAt:
          type: string
          format: date-time
          example: "2026-07-06T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-06T15:45:00Z"
        lastActivity:
          type: string
          format: date-time
          example: "2026-07-06T15:45:00Z"
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          example: "wh_456def"
        url:
          type: string
          format: uri
          example: "https://api.example.com/webhooks/flowrider"
        events:
          type: array
          items:
            type: string
          example: ["session.created", "session.output"]
        active:
          type: boolean
          example: true
        secret:
          type: string
          description: Webhook secret (masked)
          example: "whsec_***************"
        deliveryCount:
          type: integer
          description: Total number of deliveries
          example: 142
        lastDeliveryAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-07-06T15:30:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2026-07-01T12:00:00Z"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request parameters"
        code:
          type: string
          description: Error code
          example: "INVALID_INPUT"
        details:
          type: object
          nullable: true
          description: Additional error details
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid request parameters"
            code: "INVALID_INPUT"
    UnauthorizedError:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid or missing API key"
            code: "UNAUTHORIZED"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "An internal server error occurred"
            code: "INTERNAL_ERROR"